Corrected procedures structure
Accidentally forgot the initial parent table to join to proceduregroup.

Also unintentionally left the proper formatting for NEMSIS of "eProcedures" but in this instance the tables are named "eprocedures" so the hierarchical casing was corrected.

added a debug function to assist in missing tables when using the create_views script.